cont

+++ Entities/Concrete/Car.cs
@@ -15,7 +15,7 @@ namespace Entities.Concrete
 
         public string ModelYear { get; set; }
 
-        public double DailyPrice { get; set; }
+        public float DailyPrice { get; set; }
 
         public string Description { get; set; }
